+++ engine/game/game_object.hpp
@@ -17,7 +17,7 @@ class GameObject : public Object
 {
  public:
    // Constructors and destructors
-   GameObject(Application& application, const glm::vec2& position, const glm::ivec2& size,
+   GameObject(Application& application, const glm::vec2& position, const glm::vec2& size,
               const std::string& sprite, Object::TYPE type);
 
    virtual void Hit(int32_t /*dmg*/)
